Add React Router 4.2.2
Add React Router v4.2.2 and create the simple pages for the main rout
es.

Main routes:
	- Inicio
	- Convocatoria
	- Material
	- Anuncios
	- Contacto
	- Resultados

Install npm package react-router-dom, add the BrowserRouter component
at the outest level of the app. Create links with the NavLink
component and use Switch with Route components to render the correct
components to the page.